[fix]不要なhelperを削除
