--- app/helpers/accounts_helper.rb
@@ -1,2 +1,0 @@
-module AccountsHelper
-end
