--- app/helpers/scores_helper.rb
@@ -1,2 +1,0 @@
-module ScoresHelper
-end
