--- app/helpers/tests/selections_helper.rb
@@ -1,2 +1,0 @@
-module Tests::SelectionsHelper
-end
