--- app/helpers/top_helper.rb
@@ -1,2 +1,0 @@
-module TopHelper
-end
