--- app/helpers/user_responses_helper.rb
@@ -1,2 +1,0 @@
-module UserResponsesHelper
-end
