+++ spec/decorators/test_decorator_spec.rb
@@ -16,16 +16,6 @@ RSpec.describe TestDecorator do
       expect(test.decorate.implementation_year).to eq '第58回(2023年度)'
     end
   end
-  
-  describe '#question_code' do
-    let(:test){ create(:test, year:2023) }
-    let(:test_session){ create(:test_session, session: 'AM', test: ) }
-    let(:question){ create(:question, question_number: 100, test_session:)}
-
-    it '簡略化された回次・セッション・問題番号を返す' do
-      expect(test.decorate.question_code(question)).to eq '58A100'
-    end
-  end
 
   describe '#question_code_without_turn' do
     let(:test){ create(:test, year:2023) }
